+++ test-packages/test-services/openapi/test-service/openapi/api.js
@@ -24,4 +24,8 @@ var __exportStar = (this && this.__exportStar) || function(m, exports) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 __exportStar(require("./api/default-api"), exports);
+__exportStar(require("./api/entity-api"), exports);
+__exportStar(require("./api/tag-dot-api"), exports);
+__exportStar(require("./api/tag-space-api"), exports);
+__exportStar(require("./api/test-case-api"), exports);
 //# sourceMappingURL=api.js.map