+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ## New Functionality
 
+- [openapi-generator] Generate multiple API objects based on tags instead of one API per API specification.
 - [openapi-generator] Add option `writeReadme` for generating readme files.
 - [openapi-generator] Support copying files from given paths via a new option `additionalFiles`.
 - [eslint-config] Release a recommended, shareable `eslint` configuration based on the SAP Cloud SDK's internal configuration as `@sap-cloud-sdk/eslint-config`. Follow the [official guide](https://eslint.org/docs/developer-guide/shareable-configs#using-a-shareable-config) for information how to use the config.
@@ -39,7 +40,7 @@ Blog: TBD<br>
 
 ## New Functionality
 
-- [openApi-generator] Support the generation of transpiled JavaScript OpenAPI clients with the new flags: `generateJs` and `tsConfig`.
+- [openapi-generator] Support the generation of transpiled JavaScript OpenAPI clients with the new flags: `generateJs` and `tsConfig`.
 
 ## Fixed Issues